+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Audit information is now logged into a file named `audit.log` instead of inside the standard `event.log`.
+  [jochum]
+
 - Ignore errors caused by subscribers trying to access nonexistent registry records when package is still not installed (fixes `#1`_).
   [hvelarde]